fix size_t to unsigned conversion warnings
From: David Flynn <davidf@rd.bbc.co.uk>

git-svn-id: svn://hevc.kw.bbc.co.uk/svn/jctvc-tmuc/branches/HM-3.0-dev@939 ce97f6c2-8f86-0410-b44d-af84b13ec34f

+++ source/App/TAppCommon/program_options_lite.cpp
@@ -168,7 +168,7 @@ namespace df
           cout << line.str() << endl;
           continue;
         }
-        size_t currlength = line.tellp();
+        size_t currlength = size_t(line.tellp());
         if (currlength > opt_width)
         {
           /* if option text is too long (and would collide with the
